+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+### Added
+
+- Add `cursor_pagination/1` component.
+
+### Changed
+
+- The pagination component adds the `disabled` class to the `span` that is
+  displayed when the previous or next link is disabled. Previously, the
+  `disabled` attribute was set on the `span`. The class can be customized with
+  the `:disabled_class` option.
+
 ## [0.12.0] - 2021-11-08
 
 ### Added